+++ NEWS.md
@@ -2,6 +2,7 @@
 
 #### NEW FEATURES
 * The package no longer depends on `RSAGA`, making it lighter.
+* _Robin Hood_ detrending has been added.
 
 
 ### 0.4.0